+++ include/iot/container.h
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2019-2020 IOTech Ltd
+// Copyright (c) 2019-2022 IOTech Ltd
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -134,12 +134,22 @@ extern iot_component_t * iot_container_find_component (iot_container_t * cont, c
 extern void iot_container_delete_component (iot_container_t * cont, const char * name);
 
 /**
- * @brief List the components within a container
+ * @brief List state of all components
+ *
+ * @param cont     Pointer to a container
+ * @param category Category name of component to list. If NULL all components are listed
+ * @return         List of maps of component data, with keys "name", "type", "state" and "config"
+ */
+extern iot_data_t * iot_container_list_components (iot_container_t * cont, const char * category);
+
+/**
+ * @brief Get state of named component
  *
  * @param cont  Pointer to a container
- * @return      String map of component info, keyed by component name
+ * @param name  Component name
+ * @return      Data map, with keys "name", "type", "state" and "config"
  */
-extern iot_data_t * iot_container_list_components (iot_container_t * cont);
+extern iot_data_t * iot_container_component_read (iot_container_t * cont, const char * name);
 
 #ifdef __cplusplus
 }