+++ src/c/component.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020
+ * Copyright (c) 2020-2022
  * IoTech Ltd
  *
  * SPDX-License-Identifier: Apache-2.0
@@ -25,6 +25,12 @@ void iot_component_init (iot_component_t * component, const iot_component_factor
   atomic_store (&component->refs, 1u);
 }
 
+void iot_component_set_running_callback (iot_component_t * component, iot_component_running_fn_t fn)
+{
+  assert (component && fn);
+  component->running_fn = fn;
+}
+
 bool iot_component_reconfig (iot_component_t * component, iot_container_t * cont, const iot_data_t * map)
 {
   assert (component && cont && map);
@@ -34,6 +40,7 @@ bool iot_component_reconfig (iot_component_t * component, iot_container_t * cont
 void iot_component_fini (iot_component_t * component)
 {
   free (component->name);
+  iot_data_free (component->config);
   pthread_cond_destroy (&component->cond);
   pthread_mutex_destroy (&component->mutex);
 }
@@ -121,6 +128,20 @@ bool iot_component_set_deleted (iot_component_t * component)
 bool iot_component_set_starting (iot_component_t * component)
 {
   return iot_component_set_state (component, IOT_COMPONENT_STARTING);
+}
+
+extern iot_data_t * iot_component_read (iot_component_t * component)
+{
+  assert (component);
+  iot_data_t * data = iot_data_alloc_map (IOT_DATA_STRING);
+  pthread_mutex_lock (&component->mutex);
+  iot_data_map_add (data, IOT_DATA_STATIC (iot_data_consts.name), iot_data_alloc_string (component->name, IOT_DATA_REF));
+  iot_data_map_add (data, IOT_DATA_STATIC (iot_data_consts.type), iot_data_alloc_string (component->factory->type, IOT_DATA_REF));
+  iot_data_map_add (data, IOT_DATA_STATIC (iot_data_consts.state), iot_data_alloc_string (iot_component_state_name (component->state), IOT_DATA_REF));
+  iot_data_map_add (data, IOT_DATA_STATIC (iot_data_consts.config), iot_data_add_ref (component->config));
+  if (component->factory->category) iot_data_map_add (data, IOT_DATA_STATIC (iot_data_consts.category), iot_data_alloc_string (component->factory->category, IOT_DATA_REF));
+  pthread_mutex_unlock (&component->mutex);
+  return data;
 }
 
 extern const char * iot_component_state_name (iot_component_state_t state)